refactor: group custom fields components and extract shared error mapper
Move CustomFields-related components into a dedicated subfolder and
share the toFieldErrors helper through a new customFieldErrors util to
remove the duplicated implementation.
